feat: make contact page responsive and use a members file to load info
