+++ src/screens/About/Member/index.js
@@ -1,0 +1,2 @@
+import Member from './Member';
+export default Member;
