+++ src/screens/About/members.js
@@ -1,0 +1,39 @@
+const members = [
+  {
+    login: 'camiloei',
+    htmlUrl: 'https://github.com/camiloei',
+    name: 'Camilo',
+    location: 'Seattle, WA',
+    blog: '',
+  },
+  {
+    login: 'cagodoy',
+    htmlUrl: 'https://github.com/cagodoy',
+    name: 'Camilo Acuña Godoy',
+    blog: '',
+    location: 'Santiago de Chile',
+  },
+  {
+    login: 'sebastianvera',
+    htmlUrl: 'https://github.com/sebastianvera',
+    name: 'Sebastian Vera',
+    location: 'Berlin, Germany',
+    blog: 'https://sebastianvera.github.io/web/',
+  },
+  {
+    login: 'RafaelVidaurre',
+    htmlUrl: 'https://github.com/RafaelVidaurre',
+    name: 'Rafael Vidaurre',
+    location: 'Chile',
+    blog: 'https://vidaurre.io',
+  },
+  {
+    login: 'vinaybedre',
+    htmlUrl: 'https://github.com/vinaybedre',
+    name: 'Vinay',
+    location: null,
+    blog: '',
+  },
+];
+
+export default members;
